refactor: ProjectEditor 컴포넌트 분리
- ProjectCard 컴포넌트 분리 (헤더 + 확장/축소 컨테이너)
- sections 폴더 생성 및 섹션별 컴포넌트 분리
  - ProjectBasicInfo: 기본 정보 섹션
  - ProjectTechAndTasks: 기술 스택 & 작업 섹션
  - ProjectGrowthSection: 성장 경험 섹션
  - ProjectLinksSection: 링크 섹션
  - ProjectImagesSection: 이미지 섹션
  - ProjectDisplayOptions: 표시 옵션 섹션
- 720줄 → 140줄로 메인 컴포넌트 경량화
