+++ apps/web/src/components/editor/portfolio/sections/index.ts
@@ -1,0 +1,6 @@
+export { ProjectBasicInfo } from './ProjectBasicInfo';
+export { ProjectTechAndTasks } from './ProjectTechAndTasks';
+export { ProjectGrowthSection } from './ProjectGrowthSection';
+export { ProjectLinksSection } from './ProjectLinksSection';
+export { ProjectImagesSection } from './ProjectImagesSection';
+export { ProjectDisplayOptions } from './ProjectDisplayOptions';
